Updated new functions
Added wrapper functions for the searchPos and length functions, renamed the original searchPos getPosition, then made the new searchPos a wrapper function that eliminated the need to specify 0 as a start index. Same thing for length, simply wrapped that in another function called list_length, which eliminated the need for one extra parameter.